rtx: add denoiser api stub
